util: Use off_t instead of int for size values
Old output:
> Total Installed Size:  -1527.44 MiB

Fixed:
> Total Installed Size:  2568.56 MiB

Bug introduced in 7b8f8753b15037bf4a88126ffde8195416432c72.

+++ src/pacman/util.c
@@ -433,7 +433,7 @@ static void table_free(alpm_list_t *headers, alpm_list_t *rows)
 	alpm_list_free(rows);
 }
 
-static void add_transaction_sizes_row(alpm_list_t **rows, char *label, int size)
+static void add_transaction_sizes_row(alpm_list_t **rows, char *label, off_t size)
 {
 	alpm_list_t *row = NULL;
 	char *str;
